Consolidate enumerated option helpers into core workspace
**Migration Summary:**
- Moved `createEnumeratedOptionHelpers` from `src/cli/src/shared/enumerated-option-helpers.ts` to `src/core/src/utils/enumerated-options.ts`
- Updated 7 CLI files to import from Core instead of local shared directory
- Removed duplicate implementation (83 lines)
- Net change: +87/-92 lines (5 line reduction)

**Symbols Renamed/Moved:**
- `createEnumeratedOptionHelpers`: Now exported from `@gml-modules/core` (via Core namespace)
- All CLI imports now use: `import { Core } from "@gml-modules/core"; const { createEnumeratedOptionHelpers } = Core;`

**Files Updated:**
- `src/cli/src/cli-core/command-suite-helpers.ts` - Import from Core
- `src/cli/src/cli.ts` - Import from Core
- `src/cli/src/modules/dependencies.ts` - Removed re-export
- `src/cli/src/modules/memory/index.ts` - Import from Core
- `src/cli/src/modules/performance/suite-options.ts` - Import from Core
- `src/cli/src/modules/quality-report/test-status.ts` - Import from Core
- `src/cli/test/shared-enumerated-option-helpers.test.ts` - Import from Core
- `src/core/src/utils/enumerated-options.ts` - Added `createEnumeratedOptionHelpers` function

**Deleted:**
- `src/cli/src/shared/enumerated-option-helpers.ts` (83 lines)

**Testing:**
✅ All core tests pass (302 tests, 0 failures)
✅ All CLI tests pass (396 tests, 0 failures)
✅ TypeScript compilation successful
✅ Code formatted with Prettier
✅ Lint warnings fixed (toSorted, type safety)

**Downstream Adaptation:**
Consumers already using `@gml-modules/core` should destructure `createEnumeratedOptionHelpers` from the Core namespace. No breaking changes to public API - this is a drop-in consolidation that reduces duplication while maintaining identical functionality.
