+++ src/cli/src/modules/dependencies.ts
@@ -1,6 +1,5 @@
 export * from "../shared/skip-cli-run.js";
 export * from "../shared/ancestor-directories.js";
-export * from "../shared/enumerated-option-helpers.js";
 
 export { applyStandardCommandOptions } from "../cli-core/command-standard-options.js";
